requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
python-dotenv
pypdf
chromadb
torch
sentence-transformers
streamlit